pyproject.toml
[tool.poetry]
name = "tg-avatar"
version = "0.2.0"
description = "Script that makes your Telegram avatar as weather station"
authors = ["Andrey Bibea <andrebibea@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
pillow = "10.4.0"
telethon = "1.36.0"
aiohttp = "3.10.5"
pydantic = "2.9.0"
moviepy = "1.0.3"
pysocks = "1.7.1"
uvloop = "0.20.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
